backend/requirements.txt
fastapi[standard]>=0.115.0
anthropic==0.76.0
sqlalchemy>=2.0.35
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
authlib>=1.3.0
cryptography>=43.0.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
httpx>=0.27.0
